requirements.txt
openai
python-dotenv
tqdm
beautifulsoup4
requests
pandas